changed library to use float64 wherever possible, in prevision of future changes

+++ pyslic3r/_TriangleMesh.pxd
@@ -7,4 +7,4 @@ cimport slic3r_defs as s3
 cdef class TriangleMesh:
   """class to represent a STL mesh and slice it"""
   cdef s3._TriangleMesh *thisptr
-  cpdef cnp.ndarray[cnp.float32_t, ndim=2] boundingBox(self)
+  cpdef cnp.ndarray[cnp.float64_t, ndim=2] boundingBox(self)
